requirements.txt
django>=4.2
channels>=4.0
deepgram-sdk>=4.8.1
python-dotenv>=1.0.0
django-cors-headers>=4.0
daphne>=4.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-django>=4.5.0
